Fix: don't report missing ui settings file
