+++ tmaze/src/settings/config_utils.rs
@@ -143,3 +143,55 @@ pub enum Value {
     Bool(bool),
     String(String),
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_value_deserialize() {
+        let json_data = r#"
+        {
+            "name": "Example",
+            "enabled": true,
+            "threshold": 10.5,
+            "count": 42,
+            "items": [1, 2, 3],
+            "settings": {
+                "option1": "value1",
+                "option2": false
+            }
+        }
+        "#;
+
+        let parsed: Value = serde_json::from_str(json_data).unwrap();
+
+        if let Value::Object(map) = parsed {
+            assert_eq!(map.get("name"), Some(&Value::String("Example".to_string())));
+            assert_eq!(map.get("enabled"), Some(&Value::Bool(true)));
+            assert_eq!(map.get("threshold"), Some(&Value::Float(10.5)));
+            assert_eq!(map.get("count"), Some(&Value::Int(42)));
+
+            if let Some(Value::List(items)) = map.get("items") {
+                assert_eq!(items.len(), 3);
+                assert_eq!(items[0], Value::Int(1));
+                assert_eq!(items[1], Value::Int(2));
+                assert_eq!(items[2], Value::Int(3));
+            } else {
+                panic!("Expected 'items' to be a list");
+            }
+
+            if let Some(Value::Object(settings)) = map.get("settings") {
+                assert_eq!(
+                    settings.get("option1"),
+                    Some(&Value::String("value1".to_string()))
+                );
+                assert_eq!(settings.get("option2"), Some(&Value::Bool(false)));
+            } else {
+                panic!("Expected 'settings' to be an object");
+            }
+        } else {
+            panic!("Expected top-level value to be an object");
+        }
+    }
+}
